Incomplete: Event Rewrite
INCOMPLETE: The entire SudokuPanel's event system is being rebuilt to improve code navigate, and to make it easier to work with. The new event system is also more responsive then the old one.

The concept of active cell has been replaced with a new method which I believe is superior. The equivalence of active cell is not the head of the cell selection. It is now possible to have no cells selected! I felt that a constant cell selector on the screen was sometimes distracting.

I have been breaking down large functions into smaller more reusable functions, which also improves readability of the code. For instance: isOnGrid and calculateGridRegion.

Drag to select and arrow keys to control the selection head have been re-written from scratch; however, there remain many features to re-implement which I have removed when rebuilding this entire new system. It will take some time before the program is as functional as the previous pre-release.