Cell Selector
I am rewriting the main cell selector so that it is not to be confused anymore with the rest of the cell selection. I'm sure this is the design Bernhard Hobiger had in mind given the structure of the original code. It should not be possible to click outside the grid to hide the selector, a feature which I had been looking forward to for a long time. This transition is very experimental and may still have bugs left to iron out. I have not yet tested this new behaviour with Single Click Mode.