Mouse and Key Input Bug Fixes
Input digit on a cell with a user input value now overwrites it.

The Modify Givens button is not active when the program loads, and Play Game is not clickable.

Right click a user input cell shows the right click menu with the option to delete value.

Cells now color even if you drag and release on the same cell.