Active Cell Removes
Internal redesign of the active cell. Instead of holding the index of the active cell selector, we not rely on the last cell added to our cell selection.

A few mouse events have been updates, slightly more responsive. There may be a few bugs to iron out still because it was a decent change nonetheless.