Commenting out test code
The original author used mains in MANY classes to test them which is annoying when I am focused on a file in eclipse because it tries to compile this file's main which lead to wrong JAR entry point thus console functionality was not working. To address this, I commented out ALL other mains.