Active Cell Toggle Candidates
Previously, it was not possible to toggle a cell's candidate state using the Active Cell Panel. The existing feature only provided Remove Candidates which has since been repurposed in a more general manner as Toggle Candidates. This feature also works for when Show All Candidates is off.

I also refactored the construction of the UI buttons in a more compact and neater way.